menu access and menu add edit module on company panel

+++ database/migrations/2024_12_13_150322_create_menus_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string("icon");
             $table->unsignedBigInteger("parent_id")->nullable();
             $table->boolean('status')->default(true);
-            $table->unsignedBigInteger('order_no');
+            $table->integer('order_no');
             $table->timestamps();
         });
     }
